backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
pymongo[srv]
python-dotenv
Werkzeug